requirements.txt
livekit>=1.0
livekit-agents~=1.0
livekit-plugins-openai~=1.0
livekit-plugins-cartesia~=1.0
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-silero
livekit-plugins-turn-detector
torch
huggingface_hub
python-dotenv~=1.0
